+++ src/features/postComments/ui/PostComments.tsx
@@ -1,0 +1,210 @@
+import { Dispatch, SetStateAction, useState } from "react"
+import { Button, Dialog, DialogContent, DialogHeader, DialogTitle, highlightText, Textarea } from "@/shared/ui"
+import { Edit2, Plus, ThumbsUp, Trash2 } from "lucide-react"
+import { Comment, Comments } from "@/widgets/post/ui/PostWidget.tsx"
+import { PostContent } from "@/entities/post"
+
+const addCommentAPI = async (comment: { body: string; postId: number | null; userId: number }): Promise<Comment> => {
+  const response = await fetch("/api/comments/add", {
+    method: "POST",
+    headers: { "Content-Type": "application/json" },
+    body: JSON.stringify(comment),
+  })
+  return response.json()
+}
+
+const updateCommentAPI = async (comment: Comment): Promise<Comment> => {
+  const response = await fetch(`/api/comments/${comment.id}`, {
+    method: "PUT",
+    headers: { "Content-Type": "application/json" },
+    body: JSON.stringify({ body: comment.body }),
+  })
+  return response.json()
+}
+
+const deleteCommentAPI = async (id: number): Promise<void> => {
+  await fetch(`/api/comments/${id}`, {
+    method: "DELETE",
+  })
+}
+
+const likeCommentAPI = async (id: number, likes: number): Promise<Comment> => {
+  const response = await fetch(`/api/comments/${id}`, {
+    method: "PATCH",
+    headers: { "Content-Type": "application/json" },
+    body: JSON.stringify({ likes: likes + 1 }),
+  })
+  return response.json()
+}
+
+export interface NewComment {
+  body: string
+  postId: number | null
+  userId: number
+}
+
+const defaultNewComment: NewComment = {
+  body: "",
+  postId: null,
+  userId: 1,
+}
+interface Props {
+  postId: PostContent["id"]
+  comments: Comments
+  setComments: Dispatch<SetStateAction<Comments>>
+  searchQuery: string
+}
+export const PostComments = ({ postId, comments, setComments, searchQuery }: Props) => {
+  const [selectedComment, setSelectedComment] = useState<Comment | null>(null)
+  const [newComment, setNewComment] = useState<NewComment>(defaultNewComment)
+
+  const [showAddCommentDialog, setShowAddCommentDialog] = useState<boolean>(false)
+  const [showEditCommentDialog, setShowEditCommentDialog] = useState<boolean>(false)
+  const updateCommentState = (postId: number, updatedComment: Comment) => {
+    setComments((prev) => ({
+      ...prev,
+      [postId]: prev[postId].map((comment) => (comment.id === updatedComment.id ? updatedComment : comment)),
+    }))
+  }
+
+  // 댓글 좋아요
+  const likeComment = async (id: number, postId: number) => {
+    try {
+      const comment = comments[postId]?.find((c) => c.id === id)
+      if (!comment) return
+
+      const data = await likeCommentAPI(id, comment.likes)
+      updateCommentState(postId, { ...data, likes: comment.likes + 1 })
+    } catch (error) {
+      console.error("댓글 좋아요 오류:", error)
+    }
+  }
+  const deleteCommentState = (postId: number, commentId: number) => {
+    setComments((prev) => ({
+      ...prev,
+      [postId]: prev[postId].filter((comment) => comment.id !== commentId),
+    }))
+  }
+
+  // 댓글 추가
+  const addComment = async () => {
+    try {
+      const data = await addCommentAPI(newComment)
+      setComments((prev) => ({
+        ...prev,
+        [data.postId]: [...(prev[data.postId] || []), data],
+      }))
+      setShowAddCommentDialog(false)
+      setNewComment(defaultNewComment)
+    } catch (error) {
+      console.error("댓글 추가 오류:", error)
+    }
+  }
+
+  // 댓글 업데이트
+  const updateComment = async () => {
+    if (!selectedComment) return
+    try {
+      const data = await updateCommentAPI(selectedComment)
+      updateCommentState(data.postId, data)
+      setShowEditCommentDialog(false)
+    } catch (error) {
+      console.error("댓글 업데이트 오류:", error)
+    }
+  }
+
+  // 댓글 삭제
+  const deleteComment = async (id: number, postId: number) => {
+    try {
+      await deleteCommentAPI(id)
+      deleteCommentState(postId, id)
+    } catch (error) {
+      console.error("댓글 삭제 오류:", error)
+    }
+  }
+
+  return (
+    <>
+      <div className="mt-2">
+        <div className="flex items-center justify-between mb-2">
+          <h3 className="text-sm font-semibold">댓글</h3>
+          <Button
+            size="sm"
+            onClick={() => {
+              setNewComment((prev) => ({ ...prev, postId }))
+              setShowAddCommentDialog(true)
+            }}
+          >
+            <Plus className="w-3 h-3 mr-1" />
+            댓글 추가
+          </Button>
+        </div>
+        <div className="space-y-1">
+          {comments[postId]?.map((comment) => (
+            <div key={comment.id} className="flex items-center justify-between text-sm border-b pb-1">
+              <div className="flex items-center space-x-2 overflow-hidden">
+                <span className="font-medium truncate">{comment.user.username}:</span>
+                <span className="truncate">{highlightText(comment.body, searchQuery)}</span>
+              </div>
+              <div className="flex items-center space-x-1">
+                <Button variant="ghost" size="sm" onClick={() => likeComment(comment.id, postId)}>
+                  <ThumbsUp className="w-3 h-3" />
+                  <span className="ml-1 text-xs">{comment.likes}</span>
+                </Button>
+                <Button
+                  variant="ghost"
+                  size="sm"
+                  onClick={() => {
+                    setSelectedComment(comment)
+                    setShowEditCommentDialog(true)
+                  }}
+                >
+                  <Edit2 className="w-3 h-3" />
+                </Button>
+                <Button variant="ghost" size="sm" onClick={() => deleteComment(comment.id, postId)}>
+                  <Trash2 className="w-3 h-3" />
+                </Button>
+              </div>
+            </div>
+          ))}
+        </div>
+      </div>
+      {/* 댓글 추가 대화상자 */}
+      <Dialog open={showAddCommentDialog} onOpenChange={setShowAddCommentDialog}>
+        <DialogContent>
+          <DialogHeader>
+            <DialogTitle>새 댓글 추가</DialogTitle>
+          </DialogHeader>
+          <div className="space-y-4">
+            <Textarea
+              placeholder="댓글 내용"
+              value={newComment.body}
+              onChange={(e) => setNewComment({ ...newComment, body: e.target.value })}
+            />
+            <Button onClick={addComment}>댓글 추가</Button>
+          </div>
+        </DialogContent>
+      </Dialog>
+
+      {/* 댓글 수정 대화상자 */}
+      <Dialog open={showEditCommentDialog} onOpenChange={setShowEditCommentDialog}>
+        <DialogContent>
+          <DialogHeader>
+            <DialogTitle>댓글 수정</DialogTitle>
+          </DialogHeader>
+          <div className="space-y-4">
+            <Textarea
+              placeholder="댓글 내용"
+              value={selectedComment?.body || ""}
+              onChange={(e) => {
+                if (!selectedComment) return
+                setSelectedComment({ ...selectedComment, body: e.target.value })
+              }}
+            />
+            <Button onClick={updateComment}>댓글 업데이트</Button>
+          </div>
+        </DialogContent>
+      </Dialog>
+    </>
+  )
+}
