+++ src/features/postDetail/ui/PostDetailButton.tsx
@@ -3,19 +3,19 @@ import { PostContent } from "@/entities/post"
 import { Dispatch, SetStateAction, useState } from "react"
 import { Button, Dialog, DialogContent, DialogHeader, DialogTitle, highlightText } from "@/shared/ui"
 import { Comment, Comments } from "@/widgets/post/ui/PostWidget.tsx"
+import { PostComments } from "@/features/postComments/ui/PostComments.tsx"
 
 interface Props {
   post: PostContent
   comments: Comments
   setComments: Dispatch<SetStateAction<Comments>>
   searchQuery: string
-  renderComments: (postId: number) => React.ReactNode
 }
 const fetchCommentsAPI = async (postId: number): Promise<{ comments: Comment[] }> => {
   const response = await fetch(`/api/comments/post/${postId}`)
   return response.json()
 }
-export const PostDetailButton = ({ post, comments, setComments, searchQuery, renderComments }: Props) => {
+export const PostDetailButton = ({ post, comments, setComments, searchQuery }: Props) => {
   const [selectedPostForDetail, setSelectedPostForDetail] = useState<PostContent | null>(null)
   const updateCommentsState = (postId: number, newComments: Comment[]) => {
     setComments((prev) => ({ ...prev, [postId]: newComments }))
@@ -48,7 +48,12 @@ export const PostDetailButton = ({ post, comments, setComments, searchQuery, ren
           </DialogHeader>
           <div className="space-y-4">
             <p>{highlightText(selectedPostForDetail?.body || "", searchQuery)}</p>
-            {renderComments(selectedPostForDetail?.id || 0)}
+            <PostComments
+              postId={selectedPostForDetail?.id || 0}
+              setComments={setComments}
+              comments={comments}
+              searchQuery={searchQuery}
+            />
           </div>
         </DialogContent>
       </Dialog>
